+++ lib/src/links/api_link.dart
@@ -131,7 +131,7 @@ abstract class ApiLink {
   /// Calling `super.next` will cause invocation of `next` method
   /// in the next ApiLink in the chain (if present)
   @protected
-  Future<ApiResponse> next(ApiRequest request) {
+  Future<Response> next(Request request) {
     return _nextLink?.next(request);
   }