+++ lib/src/links/debug_link.dart
@@ -3,7 +3,7 @@ part of http_api;
 /// To create own dev-only ApiLink, extend this class.
 /// {@template http_api.debug_link}
 ///
-/// DebugLinks are always skipped in release builds.
+/// DebugLinks will NOT be present in release builds.
 /// They are ignored by [ApiLink.chain] method.
 /// {@endtemplate}
 abstract class DebugLink extends ApiLink {}